requirements.txt
numpy
pandas
pymysql
python-dotenv
scikit-learn
#-e .